Tableau Dashboard of "British Airways Review"
This is an example of a listing of reviews for British Airways from 2016 to 2023. The dashboard allows the user to select from a variety of filters to change the view of the pane dependent on what they would like to know. The types of ratings displayed in the main pane are the average overall rating of the month, the average overall rating by country, and the average overall rating by aircraft. Amongst these, there are many filters to adjust the ratings, such as the metric chosen, month and year chosen, traveler type, seat type, continent, and aircraft type. Have fun comparing the various metrics and filters to see changes in the charts and maps for some interesting insight!