King County, WA House Sales Tableau Workbook
This is a direct link to download and run the Tableau workbook for "King County, WA House Sales." The Tableau dashboard features a main pane displaying a chart for the daily average house sales price, a map displaying the price differences by neighborhood (cool colors for lower prices and warmer colors for higher prices), a graph displaying the distribution of house prices, a heat map showing the view versus condition in terms of pricing, a chart conveying the distribution of bedrooms in a home, and lastly another chart displaying the distribution of bathrooms in a home. There are custom filters to adjust the month between May, 2014 to May, 2015, selecting the time of day or month via a calendar, selecting the year built, the square foot of living space, and finally the square foot of the overall lot. 

Thank you for your time and I hope you enjoy tinkering around with it! Feedback is always welcome too!